% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earclip.R
\name{earclip_rgl}
\alias{earclip_rgl}
\title{Ear clipping triangulation}
\usage{
earclip_rgl(x, y = NULL, z = NULL, random = TRUE, plot = FALSE,
  partial = NA)
}
\description{
Ear clipping is a relatively cheap method for constrained triangulation
of polygons.
}
\details{
Originally this code was published in the `rgl` package, and written
by Duncan Murdoch.
}
\examples{
data("minimal_mesh", package = "silicate")
rbind_na <- function(x) {
## the inner head-1 is to remove the final closing/start point from simple features
head(do.call(rbind, lapply(x, function(a) rbind(head(a, -1), NA))), -1)
}
strip_sf <- function(x) {
  lapply(unlist(lapply(unclass(x), unclass), recursive = FALSE), rbind_na)
}
x <- strip_sf(minimal_mesh$geom)
lapply(x, earclip_rgl)
#ncsf <- sf::read_sf(system.file("shape/nc.shp", package="sf"))
#x <- strip_sf(ncsf$geometry)
#lapply(x, earclip_rgl)
## ## 20s in rgl, 14s with col/rowSums
#rbenchmark::benchmark(lapply(x[1:12], earclip_rgl), replications = 40)
}
\author{
Duncan Murdoch, in the rgl package
}
